+++ configs/config_across_scen_T_cmip6ng_test.py
@@ -43,8 +43,6 @@ esms = ["IPSL-CM6A-LR"]
 # emulated variables
 targs = ["tas"]
 
-reg_type = "srex"
-
 # reference period
 ref = {}