+++ mesmer/calibrate_mesmer/calibrate_mesmer.py
@@ -190,13 +190,14 @@ def _calibrate_and_draw_realisations(
         GSAT[esm] = convert_dict_to_arr(GSAT_dict[esm])
         GHFDS[esm] = convert_dict_to_arr(GHFDS_dict[esm])
 
+    if reg_type is not None:
+        warnings.warn("Passing ``reg_type`` no longer has any effect.", FutureWarning)
+
     # load in the constant files
-    reg_dict, ls, wgt_g, lon, lat = load_regs_ls_wgt_lon_lat(reg_type, lon, lat)
+    ls, wgt_g, lon, lat = load_regs_ls_wgt_lon_lat(lon=lon, lat=lat)
 
     # extract land
-    tas, reg_dict, ls = extract_land(
-        tas_g, reg_dict, wgt_g, ls, threshold_land=threshold_land
-    )
+    tas, ls = extract_land(tas_g, wgt=wgt_g, ls=ls, threshold_land=threshold_land)
 
     for esm in esms:
         LOGGER.info("Calibrating %s", esm)