+++ mesmer/utils/select.py
@@ -8,11 +8,12 @@ Functions to extract regions or time period of interest.
 
 
 import copy as copy
+import warnings
 
 import numpy as np
 
 
-def extract_land(var, reg_dict, wgt, ls, threshold_land=0.25):
+def extract_land(var, reg_dict=None, wgt=None, ls=None, threshold_land=0.25):
     """
     Extract all land grid points and area weights in regions and in land-sea mask for
     given threshold.
@@ -24,16 +25,9 @@ def extract_land(var, reg_dict, wgt, ls, threshold_land=0.25):
 
         - [esm][scen] (4d array (run, time, lat, lon) of variable)
 
-    reg_dict : dict
-        region dictionary with keys
-
-        - ["type"] (region type)
-        - ["abbrevs"] (abbreviations for regions)
-        - ["names"] (full names of regions)
-        - ["grids"] (3d array (regions, lat, lon) of subsampled region fraction)
-        - ["grid_b"] (2d array (lat, lon) of regions with each grid point being assigned
-          to a single region ("binary" grid))
-        - ["full"] (full Region object (for plotting region outlines))
+    reg_dict : dict | None
+        Deprecated. No longer has an effect (except changing the number of output
+        params).
 
     wgt : np.ndarray
         2d array (lat, lon) of weights to be used for area weighted means
@@ -54,13 +48,11 @@ def extract_land(var, reg_dict, wgt, ls, threshold_land=0.25):
         nested variable at land grid points dictionary with keys
 
         - [esm] (3d array (run, time, gp_l) of variable at land grid points)
-    reg_dict : dict
-        region dictionary with added keys
 
-        - ["gps_l"] (2d array (region, gp_l) of region fraction at land grid points)
-        - ["wgt_gps_l"] (2d array (region, gp_l) of area weights for each region on land)
-        - ["gp_b_l"] (1d array of region index at land grid points with each grid point
-          being assigned to a single region)
+    reg_dict : dict
+        Optional output (empty dict). Only returned when the input ``reg_dict`` is not
+        ``None``.
+
     ls : dict
         land sea dictionary with added keys
 
@@ -74,6 +66,13 @@ def extract_land(var, reg_dict, wgt, ls, threshold_land=0.25):
           fraction)
 
     """
+
+    if reg_dict is not None:
+        warnings.warn(
+            "Passing `reg_dict` no longer has an effect. When passing `None` this "
+            "function will only return two parameters.",
+            FutureWarning,
+        )
 
     # determine which grid points count as land grid points
     idx_l = ls["grid_no_ANT"] > threshold_land
@@ -91,19 +90,6 @@ def extract_land(var, reg_dict, wgt, ls, threshold_land=0.25):
     # weights for land points: multiply weight by land fraction
     ls["wgt_gp_l"] = wgt[idx_l] * ls["gp_l"]
 
-    # extract the land points + weights from the region grids
-    reg_dict["gps_l"] = reg_dict["grids"][:, idx_l]  # country is the first axis
-    # weights for regions (1st axis): region fraction * area weights
-    reg_dict["wgt_gps_l"] = wgt[idx_l] * reg_dict["gps_l"]
-
-    if reg_dict["type"] in ["srex", "ar6.land"]:
-        # multiply by land fraction to account for coastal cells
-        # because SREX / ar6.land regions include ocean
-        reg_dict["wgt_gps_l"] = reg_dict["wgt_gps_l"] * ls["gp_l"]
-
-    # not sure if needed; extracts land from the "binary" mask
-    reg_dict["gp_b_l"] = reg_dict["grid_b"][idx_l]
-
     # extract the land points of the variable of interest
     var_l = {}
     for esm in var.keys():
@@ -112,7 +98,10 @@ def extract_land(var, reg_dict, wgt, ls, threshold_land=0.25):
             # run is the first axis, followed by time
             var_l[esm][scen] = var[esm][scen][:, :, idx_l]
 
-    return var_l, reg_dict, ls
+    if reg_dict is None:
+        return var_l, ls
+
+    return var_l, {}, ls
 
 
 def extract_time_period(var, time, start, end):